feat:删除单位改由调用存储过程并进行事务处理

+++ src/main/java/cn/xrp/projectmanagementbackend/controller/UnitController.java
@@ -36,7 +36,7 @@ public class UnitController {
     }
     @PostMapping("/delete")
     public BaseResponse<Boolean> deleteUnit(@RequestBody JsonNode index){
-        long unitId = index.get("id").longValue();
+        Integer unitId = index.get("id").intValue();
         if(unitId < 0){
             throw new BusinessException(ErrorCode.PARAMS_ERROR);
         }
